+++ src/app/Http/Controllers/UserCrudController.php
@@ -17,9 +17,9 @@ class UserCrudController extends CrudController
 
     public function setup()
     {
-        $this->crud->setModel(config('backpack.permissionmanager.models.user'));
-        $this->crud->setEntityNameStrings(trans('backpack::permissionmanager.user'), trans('backpack::permissionmanager.users'));
-        $this->crud->setRoute(backpack_url('user'));
+        $this->crud->setModel(config('hncore.permissionmanager.models.user'));
+        $this->crud->setEntityNameStrings(trans('hncore::permissionmanager.user'), trans('hncore::permissionmanager.users'));
+        $this->crud->setRoute(hncore_url('user'));
     }
 
     public function setupListOperation()
@@ -27,16 +27,16 @@ class UserCrudController extends CrudController
         $this->crud->addColumns([
             [
                 'name'  => 'name',
-                'label' => trans('backpack::permissionmanager.name'),
+                'label' => trans('hncore::permissionmanager.name'),
                 'type'  => 'text',
             ],
             [
                 'name'  => 'email',
-                'label' => trans('backpack::permissionmanager.email'),
+                'label' => trans('hncore::permissionmanager.email'),
                 'type'  => 'email',
             ],
             [ // n-n relationship (with pivot table)
-                'label'     => trans('backpack::permissionmanager.roles'), // Table column heading
+                'label'     => trans('hncore::permissionmanager.roles'), // Table column heading
                 'type'      => 'select_multiple',
                 'name'      => 'roles', // the method that defines the relationship in your Model
                 'entity'    => 'roles', // the method that defines the relationship in your Model
@@ -44,7 +44,7 @@ class UserCrudController extends CrudController
                 'model'     => config('permission.models.role'), // foreign key model
             ],
             [ // n-n relationship (with pivot table)
-                'label'     => trans('backpack::permissionmanager.extra_permissions'), // Table column heading
+                'label'     => trans('hncore::permissionmanager.extra_permissions'), // Table column heading
                 'type'      => 'select_multiple',
                 'name'      => 'permissions', // the method that defines the relationship in your Model
                 'entity'    => 'permissions', // the method that defines the relationship in your Model
@@ -53,13 +53,13 @@ class UserCrudController extends CrudController
             ],
         ]);
 
-        if (backpack_pro()) {
+        if (hncore_pro()) {
             // Role Filter
             $this->crud->addFilter(
                 [
                     'name'  => 'role',
                     'type'  => 'dropdown',
-                    'label' => trans('backpack::permissionmanager.role'),
+                    'label' => trans('hncore::permissionmanager.role'),
                 ],
                 config('permission.models.role')::all()->pluck('name', 'id')->toArray(),
                 function ($value) { // if the filter is active
@@ -74,7 +74,7 @@ class UserCrudController extends CrudController
                 [
                     'name'  => 'permissions',
                     'type'  => 'select2',
-                    'label' => trans('backpack::permissionmanager.extra_permissions'),
+                    'label' => trans('hncore::permissionmanager.extra_permissions'),
                 ],
                 config('permission.models.permission')::all()->pluck('name', 'id')->toArray(),
                 function ($value) { // if the filter is active
@@ -105,13 +105,13 @@ class UserCrudController extends CrudController
         $this->crud->column('email');
         $this->crud->column([
             // two interconnected entities
-            'label'             => trans('backpack::permissionmanager.user_role_permission'),
+            'label'             => trans('hncore::permissionmanager.user_role_permission'),
             'field_unique_name' => 'user_role_permission',
             'type'              => 'checklist_dependency',
             'name'              => 'roles_permissions',
             'subfields'         => [
                 'primary' => [
-                    'label'            => trans('backpack::permissionmanager.role'),
+                    'label'            => trans('hncore::permissionmanager.role'),
                     'name'             => 'roles', // the method that defines the relationship in your Model
                     'entity'           => 'roles', // the method that defines the relationship in your Model
                     'entity_secondary' => 'permissions', // the method that defines the relationship in your Model
@@ -119,7 +119,7 @@ class UserCrudController extends CrudController
                     'model'            => config('permission.models.role'), // foreign key model
                 ],
                 'secondary' => [
-                    'label'            => mb_ucfirst(trans('backpack::permissionmanager.permission_singular')),
+                    'label'            => mb_ucfirst(trans('hncore::permissionmanager.permission_singular')),
                     'name'             => 'permissions', // the method that defines the relationship in your Model
                     'entity'           => 'permissions', // the method that defines the relationship in your Model
                     'entity_primary'   => 'roles', // the method that defines the relationship in your Model
@@ -185,33 +185,33 @@ class UserCrudController extends CrudController
         $this->crud->addFields([
             [
                 'name'  => 'name',
-                'label' => trans('backpack::permissionmanager.name'),
+                'label' => trans('hncore::permissionmanager.name'),
                 'type'  => 'text',
             ],
             [
                 'name'  => 'email',
-                'label' => trans('backpack::permissionmanager.email'),
+                'label' => trans('hncore::permissionmanager.email'),
                 'type'  => 'email',
             ],
             [
                 'name'  => 'password',
-                'label' => trans('backpack::permissionmanager.password'),
+                'label' => trans('hncore::permissionmanager.password'),
                 'type'  => 'password',
             ],
             [
                 'name'  => 'password_confirmation',
-                'label' => trans('backpack::permissionmanager.password_confirmation'),
+                'label' => trans('hncore::permissionmanager.password_confirmation'),
                 'type'  => 'password',
             ],
             [
                 // two interconnected entities
-                'label'             => trans('backpack::permissionmanager.user_role_permission'),
+                'label'             => trans('hncore::permissionmanager.user_role_permission'),
                 'field_unique_name' => 'user_role_permission',
                 'type'              => 'checklist_dependency',
                 'name'              => 'roles,permissions',
                 'subfields'         => [
                     'primary' => [
-                        'label'            => trans('backpack::permissionmanager.roles'),
+                        'label'            => trans('hncore::permissionmanager.roles'),
                         'name'             => 'roles', // the method that defines the relationship in your Model
                         'entity'           => 'roles', // the method that defines the relationship in your Model
                         'entity_secondary' => 'permissions', // the method that defines the relationship in your Model
@@ -221,7 +221,7 @@ class UserCrudController extends CrudController
                         'number_columns'   => 3, //can be 1,2,3,4,6
                     ],
                     'secondary' => [
-                        'label'          => mb_ucfirst(trans('backpack::permissionmanager.permission_plural')),
+                        'label'          => mb_ucfirst(trans('hncore::permissionmanager.permission_plural')),
                         'name'           => 'permissions', // the method that defines the relationship in your Model
                         'entity'         => 'permissions', // the method that defines the relationship in your Model
                         'entity_primary' => 'roles', // the method that defines the relationship in your Model