+++ src/app/Http/Requests/PermissionStoreCrudRequest.php
@@ -14,7 +14,7 @@ class PermissionStoreCrudRequest extends FormRequest
     public function authorize()
     {
         // only allow updates if the user is logged in
-        return backpack_auth()->check();
+        return hncore_auth()->check();
     }
 
     /**