+++ src/app/Http/Requests/UserStoreCrudRequest.php
@@ -14,7 +14,7 @@ class UserStoreCrudRequest extends FormRequest
     public function authorize()
     {
         // only allow updates if the user is logged in
-        return backpack_auth()->check();
+        return hncore_auth()->check();
     }
 
     /**
@@ -25,7 +25,7 @@ class UserStoreCrudRequest extends FormRequest
     public function rules()
     {
         return [
-            'email'    => 'required|unique:'.config('backpack.permissionmanager.models.user', 'users').',email',
+            'email'    => 'required|unique:'.config('hncore.permissionmanager.models.user', 'users').',email',
             'name'     => 'required',
             'password' => 'required|confirmed',
         ];