+++ src/routes/backpack/permissionmanager.php
@@ -12,8 +12,8 @@
 
 Route::group([
     'namespace'  => 'Backpack\PermissionManager\app\Http\Controllers',
-    'prefix'     => config('backpack.base.route_prefix', 'admin'),
-    'middleware' => ['web', backpack_middleware()],
+    'prefix'     => config('hncore.base.route_prefix', 'admin'),
+    'middleware' => ['web', hncore_middleware()],
 ], function () {
     Route::crud('permission', 'PermissionCrudController');
     Route::crud('role', 'RoleCrudController');